+++ src/App.js
@@ -4,8 +4,7 @@ import Dashboard from "./Pages/Dashboard";
 import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 import PrivateRoute from "./components/PrivateRoute";
 import TestLogin from "./Pages/TestLogin";
-import BottomNav from "./components/common/BottomNav";
-import { RippleButton } from "./components/common/buttons/RippleButton";
+import BottomNav from "./components/BottomNav/BottomNav";
 
 const App = () => {
   return (
@@ -17,7 +16,6 @@ const App = () => {
 
         <Route path="/">
           <TestLogin />
-          <RippleButton />
           <BottomNav />
         </Route>
       </Switch>
