+++ src/components/BottomNav/RippleButton.css
@@ -1,21 +1,15 @@
-.btn {
-  margin: 50px auto;
-  border-radius: 25px;
-  background-color: #5300e8;
-  box-shadow: 0 2px 4px 0 #888888;
+.bottomNavButton {
+  padding: 15px;
   display: inline-block;
-  padding: 15px 50px;
-  color: #ffffff;
 }
 
-.center {
+.rippleButtonDemo {
   text-align: center;
 }
 
 /* ripple */
 .ripple {
   position: relative;
-  overflow: hidden;
 }
 .ripple .rippleContainer {
   position: absolute;
@@ -29,7 +23,7 @@
   border-radius: 100%;
   position: absolute;
   opacity: 0.75;
-  background-color: #ffffff;
+  background-color: var(--text-bright);
   animation: ripple 850ms;
 }
